requirements.txt
fastapi[standard]
celery==5.3.4
redis==5.0.1
PyPDF2
pymupdf
langchain
langchain_community
faiss-cpu
langchain-huggingface
groq
langchain-groq
python-dotenv
flower==2.0.1
psycopg2-binary
loguru
pymongo
PyJWT
passlib[bcrypt]
sentence-transformers